backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
MarkupSafe==3.0.2
openai==1.54.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
Werkzeug==3.1.3
zipp==3.21.0
